pyproject.toml
[project]
name = "patres-test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "greenlet>=3.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.13",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
[tool.ruff]
exclude = ["alembic/env.py"]
